[feat]: Add setting menu. Add resize and volume changing functions.
Add setting menu
Add prototype of volume changing(music and sound)
Add resizing function

Change the observer call of MouseObserver and KeyboardObserver to reflect the position in the scaled window wrt. the size of the original window

+++ src/TestMockClasses/SoundClickTrigger.cpp
@@ -5,7 +5,7 @@
 #include "Core/UserEvent.hpp"
 #include "Utility/logger.hpp"
 void SoundClickTrigger::onLeftMousePress(const sf::Vector2f &worldPosition,
-                                         const sf::Vector2i &windowPosition) {
+                                         const sf::Vector2f &windowPosition) {
     Logger::info("Left press");
 }
 
@@ -13,21 +13,21 @@ SoundClickTrigger::SoundClickTrigger(ResourceManager &resManager)
     : resManager{resManager} {}
 
 void SoundClickTrigger::onLeftMouseRelease(const sf::Vector2f &worldPosition,
-                                           const sf::Vector2i &windowPosition) {
+                                           const sf::Vector2f &windowPosition) {
     Logger::info("Left release");
 }
 void SoundClickTrigger::onRightMousePress(const sf::Vector2f &worldPosition,
-                                          const sf::Vector2i &windowPosition) {
+                                          const sf::Vector2f &windowPosition) {
     Logger::info("Right press");
 }
 void SoundClickTrigger::onRightMouseRelease(
-    const sf::Vector2f &worldPosition, const sf::Vector2i &windowPosition) {
+    const sf::Vector2f &worldPosition, const sf::Vector2f &windowPosition) {
     Logger::info("Right release");
 }
 
 void SoundClickTrigger::onMouseEvent(Mouse button, UserEvent event,
                                      const sf::Vector2f &worldPosition,
-                                     const sf::Vector2i &windowPosition) {
+                                     const sf::Vector2f &windowPosition) {
     if (button == Mouse::Left) {
         if (event == UserEvent::Press)
             return onLeftMousePress(worldPosition, windowPosition);
@@ -45,7 +45,7 @@ void SoundClickTrigger::onMouseEvent(Mouse button, UserEvent event,
 
 void SoundClickTrigger::onKeyEvent(Key key, UserEvent event,
                                    const sf::Vector2f &worldPosition,
-                                   const sf::Vector2i &windowPosition) {
+                                   const sf::Vector2f &windowPosition) {
     Logger::info(std::to_string(static_cast<int>(key)));
     Logger::info(std::to_string(static_cast<int>(event)));
 }
